Do not die if no user logged in
Since we leave it to the Webserver to handle authentication
the admin might decide to allow access to GitWebAdmin unauthenticated
on some cases.

In this case return true in has_readable if the repository is
public (has gitweb and/or daemon enabled) to be consistent with
the gitosis config we write.